refactored db models into models.py
