+++ quickannotator/api/v1/image/__init__.py
@@ -5,7 +5,6 @@ from flask.views import MethodView
 from flask import current_app, request, send_from_directory, send_file
 from werkzeug.datastructures import FileStorage
 from datetime import datetime
-import quickannotator.db as qadb
 from quickannotator.db import db
 import openslide as ops
 import openslide.deepzoom as deepzoom
@@ -13,6 +12,8 @@ from marshmallow_sqlalchemy import SQLAlchemyAutoSchema
 import large_image
 import os
 import io
+
+import quickannotator.db.models
 # from .helper import getTile
 
 
@@ -22,9 +23,9 @@ bp = Blueprint('image', __name__, description='Image operations')
 class ImageRespSchema(SQLAlchemyAutoSchema):
     """     Image response schema      """
     class Meta:
-        model = qadb.Image
+        model = quickannotator.db.models.Image
 
-    embedding_coord = qadb.GeometryField()
+    embedding_coord = quickannotator.db.models.GeometryField()
 
 class GetImageArgsSchema(Schema):
     image_id = fields.Int(required=True)
@@ -54,9 +55,9 @@ class Image(MethodView):
         """     returns an Image
         """
         result = db.session.query(
-            *[getattr(qadb.Image, column.name) for column in qadb.Image.__table__.columns],
-            db.func.ST_AsGeoJSON(qadb.Image.embedding_coord).label('embedding_coord')
-        ).filter(qadb.Image.id == args['image_id']).first()
+            *[getattr(quickannotator.db.models.Image, column.name) for column in quickannotator.db.models.Image.__table__.columns],
+            db.func.ST_AsGeoJSON(quickannotator.db.models.Image.embedding_coord).label('embedding_coord')
+        ).filter(quickannotator.db.models.Image.id == args['image_id']).first()
         if result is not None:
             return result, 200
         else:
@@ -78,7 +79,7 @@ class Image(MethodView):
     def delete(self, args):
         """     delete an Image   """
 
-        db.session.query(qadb.Image).filter(id=args['image_id']).delete()
+        db.session.query(quickannotator.db.models.Image).filter(id=args['image_id']).delete()
         return 204
 
 #################################################################################
@@ -89,7 +90,7 @@ class ImageSearch(MethodView):
     def get(self, args, project_id):
         """     returns a list of Images
         """
-        result = db.session.query(qadb.Image).filter(qadb.Image.project_id == project_id).all()
+        result = db.session.query(quickannotator.db.models.Image).filter(quickannotator.db.models.Image.project_id == project_id).all()
         return result, 200
 
 #################################################################################
@@ -97,7 +98,7 @@ class ImageSearch(MethodView):
 class ImageFile(MethodView):
     def get(self, image_id, file_type):
         """     returns an Image file   """
-        result = db.session.query(qadb.Image).filter(qadb.Image.id == image_id).first()
+        result = db.session.query(quickannotator.db.models.Image).filter(quickannotator.db.models.Image.id == image_id).first()
 
         if file_type == 1:  # image file
             return send_from_directory(result['path'], result['name'])
@@ -117,7 +118,7 @@ class ImageFile(MethodView):
     def delete(self, image_id, file_type):
         """     delete an Image file   """
 
-        result = db.session.query(qadb.Image).filter(qadb.Image.id == image_id).first()
+        result = db.session.query(quickannotator.db.models.Image).filter(quickannotator.db.models.Image.id == image_id).first()
 
         if file_type == 1:  # image file
             # TODO implement image file deletion
@@ -134,7 +135,7 @@ class PatchFile(MethodView):
     def get(self, image_id, level, col, row, file_format):
         """     returns a patch file   """
 
-        path = qadb.Image.query.get(image_id).path
+        path = quickannotator.db.models.Image.query.get(image_id).path
         full_path = os.path.join(current_app.root_path, path)
         img = large_image.open(full_path)
