+++ quickannotator/db/utils.py
