+++ quickannotator/dl/database.py
@@ -2,7 +2,8 @@ import os
 import shutil
 from sqlalchemy import create_engine, event
 from sqlalchemy.orm import sessionmaker
-from quickannotator.db import db, Image, AnnotationClass, Tile
+from quickannotator.db import db
+from quickannotator.db.models import Annotation, AnnotationClass, Image, Notification, Project, Setting, Tile
 
 def get_database_path():
     return "sqlite:////opt/QuickAnnotator/quickannotator/instance/quickannotator.db"
@@ -46,6 +47,6 @@ import cv2
 from sqlalchemy import inspect
 
 
-from quickannotator.db import db, Project, Image, AnnotationClass, Notification, Tile, Setting, Annotation, SearchCache
+from quickannotator.db import db, SearchCache
